resolve conflict  class SyEmployeePub, SyEmployeePubImp

+++ nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.app/src/main/java/nts/uk/ctx/at/shared/app/find/yearholidaygrant/GrantConditionDto.java
@@ -20,7 +20,7 @@ public class GrantConditionDto {
 	private int conditionNo;
 	
 	/* 条件値 */
-	private Integer conditionValue;
+	private Double conditionValue;
 	
 	/* 条件利用区分 */
 	private int useConditionAtr;
